+++ src/main/java/com/iroff/supportlab/SupportLabBackendApplication.java
@@ -1,4 +1,4 @@
-package com.iroff.support_lab_backend;
+package com.iroff.supportlab;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
